showCrime/requirements.txt
boxsdk==2.0.0
Django==1.11.18
django-environ==0.4.5
django-postgres-copy==2.3.7
djangorestframework==3.9.0
editdistance==0.4
geojson==2.4.1
gevent==1.4.0
googlemaps==3.0.2
gunicorn==19.9.0
mapbox==0.17.2
matplotlib==2.2.3
numpy==1.15.4
psycopg2-binary==2.7.6.1
python-dateutil==2.7.5
pytz==2018.7
sodapy==1.5.0
